requirements.txt
comtypes>=1.1.14
PyAudio>=0.2.12
pypiwin32>=223
pyttsx3>=2.90
pywin32>=304
SpeechRecognition>=3.8.1
PyAutoGUI>=0.9.53
python-decouple>=3.6


requests>=2.28.1
wikipedia>=1.4.0
pywhatkit>=5.4